Sidebar: use position sticky when available
This gets rid of the jittery effect when scrolling with the mouse wheel.
Unfortunately, it also makes it impossible to scroll to the rest of the
sidebar if it vertically overflows the viewport.

So the choice is yours: either use an older browser which doesn't
support position: sticky, and the sidebar might get jittery, but you'll
be able to scroll it (if somewhat inconveniently, see previous commit
message). Or you can use a newer browser, but then you can't scroll the
sidebar at all. It sticks to the top, which achieves the desired goal if
it's short enough, but if the sidebar content overflows the viewport,
there's no way to get to it.

Patch taken from: <https://github.com/python/cpython/commit/8ab24b2ebcf99f703f00297cb3a0c3ff857eecf8>